requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR and Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
PyMuPDF==1.23.9
numpy==1.24.4

# LLM dependencies
openai==1.3.0
google-generativeai==0.3.0

# File handling and validation
python-magic-bin==0.4.14
aiofiles==23.2.0

# HTTP client
aiohttp==3.9.0
httpx==0.25.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Deployment
python-dotenv==1.0.0
